refactor: move input/queue into trait for playlist
