+++ app/Models/Game.php
@@ -3,8 +3,6 @@
 namespace App\Models;
 
 use App\Enums\Experience;
-use App\Enums\Input;
-use App\Enums\Queue;
 use App\Jobs\PullAppearance;
 use App\Models\Contracts\HasHaloDotApi;
 use App\Models\Pivots\PersonalResult;
@@ -25,18 +23,16 @@ use Illuminate\Support\Arr;
  * @property string $uuid
  * @property int $category_id
  * @property int $map_id
- * @property boolean $is_ranked
  * @property boolean $is_ffa
  * @property boolean $is_scored
  * @property Experience $experience
- * @property Queue $queue
- * @property Input $input
  * @property Carbon $occurred_at
  * @property int $duration_seconds
  * @property string $version
  * @property boolean $was_pulled
  * @property-read Category $category
  * @property-read Map $map
+ * @property-read Playlist $playlist
  * @property-read PersonalResult $personal
  * @property-read GamePlayer[]|Collection $players
  * @property-read GameTeam[]|Collection $teams
@@ -52,7 +48,8 @@ class Game extends Model implements HasHaloDotApi
     public $guarded = [
         'id',
         'category_id',
-        'map_id'
+        'map_id',
+        'playlist_id',
     ];
 
     public $dates = [
@@ -60,14 +57,13 @@ class Game extends Model implements HasHaloDotApi
     ];
 
     public $casts = [
-        'experience' => Experience::class,
-        'queue' => Queue::class,
-        'input' => Input::class,
+        'experience' => Experience::class
     ];
 
     public $with = [
         'category',
-        'map'
+        'map',
+        'playlist',
     ];
 
     public $timestamps = false;
@@ -85,34 +81,6 @@ class Game extends Model implements HasHaloDotApi
         }
 
         $this->attributes['experience'] = $experience->value;
-    }
-
-    public function setQueueAttribute(?string $value): void
-    {
-        if (is_null($value)) {
-            return;
-        }
-
-        $queue = is_numeric($value) ? Queue::fromValue((int) $value) : Queue::coerce($value);
-        if (empty($queue)) {
-            throw new \InvalidArgumentException('Invalid Queue Enum (' . $value . ')');
-        }
-
-        $this->attributes['queue'] = $queue->value;
-    }
-
-    public function setInputAttribute(?string $value): void
-    {
-        if (is_null($value)) {
-            return;
-        }
-
-        $input = is_numeric($value) ? Input::fromValue((int) $value) : Input::coerce($value);
-        if (empty($input)) {
-            throw new \InvalidArgumentException('Invalid Input Enum (' . $value . ')');
-        }
-
-        $this->attributes['input'] = $input->value;
     }
 
     public function getNameAttribute(): string
@@ -155,6 +123,7 @@ class Game extends Model implements HasHaloDotApi
         $gameId = Arr::get($payload, 'id');
         $category = Category::fromHaloDotApi(Arr::get($payload, 'details.category'));
         $map = Map::fromHaloDotApi(Arr::get($payload, 'details.map'));
+        $playlist = Playlist::fromHaloDotApi(Arr::get($payload, 'details.playlist'));
 
         /** @var Game $game */
         $game = self::query()
@@ -165,12 +134,10 @@ class Game extends Model implements HasHaloDotApi
 
         $game->category()->associate($category);
         $game->map()->associate($map);
-        $game->is_ranked = (bool) Arr::get($payload, 'details.playlist.ranked');
+        $game->playlist()->associate($playlist);
         $game->is_ffa = !(bool) Arr::get($payload, 'teams.enabled');
         $game->is_scored = (bool) Arr::get($payload, 'teams.scoring');
         $game->experience = Arr::get($payload, 'experience');
-        $game->queue = Arr::get($payload, 'details.playlist.queue');
-        $game->input = Arr::get($payload, 'details.playlist.input');
         $game->occurred_at = Arr::get($payload, 'played_at');
         $game->duration_seconds = Arr::get($payload, 'duration.seconds');
         $game->version = config('services.autocode.version');
@@ -219,6 +186,11 @@ class Game extends Model implements HasHaloDotApi
     public function map(): BelongsTo
     {
         return $this->belongsTo(Map::class);
+    }
+
+    public function playlist(): BelongsTo
+    {
+        return $this->belongsTo(Playlist::class);
     }
 
     public function players(): HasMany
