+++ app/Models/Traits/HasPlaylist.php
@@ -27,4 +27,32 @@ trait HasPlaylist
 
         return null;
     }
+
+    public function setQueueAttribute(?string $value): void
+    {
+        if (is_null($value)) {
+            return;
+        }
+
+        $queue = is_numeric($value) ? Queue::fromValue((int) $value) : Queue::coerce($value);
+        if (empty($queue)) {
+            throw new \InvalidArgumentException('Invalid Queue Enum (' . $value . ')');
+        }
+
+        $this->attributes['queue'] = $queue->value;
+    }
+
+    public function setInputAttribute(?string $value): void
+    {
+        if (is_null($value)) {
+            return;
+        }
+
+        $input = is_numeric($value) ? Input::fromValue((int) $value) : Input::coerce($value);
+        if (empty($input)) {
+            throw new \InvalidArgumentException('Invalid Input Enum (' . $value . ')');
+        }
+
+        $this->attributes['input'] = $input->value;
+    }
 }
